+++ tests/src/Functional/ShopAdminTest.php
@@ -20,7 +20,7 @@ class ShopAdminTest extends CommerceBrowserTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'commerce_trustedshops',
   ];