+++ tests/src/Kernel/API/APITestBase.php
@@ -44,7 +44,7 @@ abstract class APITestBase extends CommerceKernelTestBase {
   /**
    * {@inheritdoc}
    */
-  public static $modules = [
+  protected static $modules = [
     'entity_reference_revisions',
     'state_machine',
     'profile',