+++ tests/src/Kernel/DefaultOrderLanguageResolverTest.php
@@ -66,7 +66,7 @@ class DefaultOrderLanguageResolverTest extends DrupalCommerceKernelTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'locale',
     'language',
     'content_translation',