+++ tests/src/Kernel/Entity/ShopTest.php
@@ -29,7 +29,7 @@ class ShopTest extends CommerceKernelTestBase {
    *
    * @var array
    */
-  public static $modules = [
+  protected static $modules = [
     'commerce_trustedshops',
   ];