+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix deprecated uncalling accessCheck relying on entity queries to check access
 - fix usage of deprecated Symfony\Component\EventDispatcher\Event in favor of Symfony\Contracts\EventDispatcher\Event
 - fix deprecation of theme classy for Drupal 10 compatibilities
+- fix tests self::modules property must be declared protected
 
 ## [2.0.0] - 2023-01-27
 ### Fixed